fix can't get data when remove file

+++ FinancialPlanningBE/FinancialPlanning.WebAPI/Controllers/ReportController.cs
@@ -85,49 +85,65 @@ namespace FinancialPlanning.WebAPI.Controllers
         {
             try
             {
+                List<Expense> expenses;
+                //Get report
                 var report = await _reportService.GetReportById(id);
+                //Get reportVersions
                 var reportVersions = await _reportService.GetReportVersionsAsync(id);
                 //string reportName = report.ReportName;
                 string reportName = "CorrectPlan";
+                //Get url of file on cloud
                 string url = await _reportService.GetFileByName(reportName + ".xlsx");
-                var savePath = Path.Combine(Directory.GetCurrentDirectory(), "Upload\\Files", reportName + ".xlsx");
 
+                //If folder doesn't exit -> create
+                var directoryPath = Path.Combine(Directory.GetCurrentDirectory(), "Upload\\Files");
+                if (!Directory.Exists(directoryPath))
+                {
+                    Directory.CreateDirectory(directoryPath);
+                }
+
+                var savePath = Path.Combine(directoryPath, reportName + ".xlsx");
+                
                 bool isDownLoad = await _fileService.DownloadFile(url, savePath);
                 //download file sucessfull 
                 if (isDownLoad)
                 {
                     // conver file to list expense
-                    var fileStream = new FileStream(savePath, FileMode.Open, FileAccess.Read);
-
-                    try
+                    using (var fileStream = new FileStream(savePath, FileMode.OpenOrCreate, FileAccess.Read))
                     {
-                        List<Expense> expenses = _fileService.ConvertExcelToList(fileStream, 0);
-                        //mapper
-                        var reportViewModel = _mapper.Map<ReportViewModel>(report);
-                        var reportVersionModel = _mapper.Map<IEnumerable<ReportVersionModel>>(reportVersions);
-                        // Get the name of the user who uploaded the file
-                        var firstReportVersion = reportVersionModel.FirstOrDefault();
-                        var uploadedBy = firstReportVersion != null ? firstReportVersion.UploadedBy : null;
-                        //remove file after download
-                        fileStream.Close();
-                        System.IO.File.Delete(savePath);
-
-                        var result = new
+                        try
                         {
-                            Report = reportViewModel,
-                            Expenses = expenses,
-                            ReportVersions = reportVersionModel,
-                            UploadedBy = uploadedBy
-                        };
-
-                        return Ok(result);
+                            //change 1 when have file
+                            expenses = _fileService.ConvertExcelToList(fileStream, 0);
+                        }
+                        catch
+                        {
+                            return BadRequest("Failed to convert");
+                        }
                     }
-                    catch
+                    //mapper
+                    var reportViewModel = _mapper.Map<ReportViewModel>(report);
+                    var reportVersionModel = _mapper.Map<IEnumerable<ReportVersionModel>>(reportVersions);
+                    // Get the name of the user who uploaded the file
+                    var firstReportVersion = reportVersionModel.FirstOrDefault();
+                    var uploadedBy = firstReportVersion != null ? firstReportVersion.UploadedBy : null;
+                    //remove file
+                    if (System.IO.File.Exists(savePath))
                     {
-                        return BadRequest("Failed to convert");
+                        System.IO.File.Delete(savePath);
                     }
-                }
 
+                    var result = new
+                    {
+                        Report = reportViewModel,
+                        Expenses = expenses,
+                        ReportVersions = reportVersionModel,
+                        UploadedBy = uploadedBy
+                    };
+                    
+                    return Ok(result);
+
+                }
                 else
                 {
                     return BadRequest("Failed to download file");
@@ -139,6 +155,7 @@ namespace FinancialPlanning.WebAPI.Controllers
                 return StatusCode(500, $"Error : {ex.Message}");
             }
         }
+
         //export report 
         [HttpGet("export/{id:guid}/{version:int}")]
         //[Authorize(Roles = "Accountant, FinancialStaff")]
